search function done

+++ list.php
@@ -20,27 +20,23 @@ include("includes/config.php");
         ?>
     </header>
     <main>
-        <div class="form-createpost">
-    
-        </div>
 
         <table class="form">
-            <tr>
+            <tr class="form">
                 <th>Produkt:</th>
                 <th>Mängd:</th>
                 <th>Plats:</th>
-                <th>Tid:</th>
                 <th>Del:</th>
             </tr>
 
             <?php
-        
+
             //instans av klassen
             $newinventory = new Newinventory();
 
-             // raderar post
-             if (isset($_GET['delete'])) {
-                $id_payment = $_GET['delete'];
+            // raderar post
+            if (isset($_GET['delete'])) {
+                $id = $_GET['delete'];
 
                 if ($newinventory->deleteProduct($id)) {
                 }
@@ -54,20 +50,34 @@ include("includes/config.php");
             // print_r($list);
             // echo "</pre>";
 
-            foreach ($list as $inventory) {
-                echo "
-                    <tr>
-                        <td>" . $inventory['product'] . "</td>
-                        <td>" . $inventory['amount'] .  "</td>
-                        <td>" . $inventory['sort'] . "</td>
-                        <td>" . $inventory['created'] . "</td>
-                        <td> <a href='list.php?delete=" . $inventory['id'] . "'> <i class='fa-solid fa-minus'></i></a> </td>
-                    </tr>
-                ";
+            if (count($list) === 0) {
+                $_SESSION['msg'] = "<h4> Listan är tom </h4>";
+            } else {
+                foreach ($list as $inventory) {
+                    echo "
+                        <tr class='form'>
+                            <td>" . $inventory['product'] . "</td>
+                            <td>" . $inventory['amount'] .  "</td>
+                            <td>" . $inventory['sort'] . "</td>
+                            <td> <a href='list.php?delete=" . $inventory['id'] . "'> <i class='fa-solid fa-minus'></i></a> </td>
+                        </tr>
+                    ";
+                }
             }
+
+
             ?>
 
         </table>
+        <div class="form">
+            <?php
+            // print all error messeges if success not true
+            if (isset($_SESSION['msg'])) {
+                echo $_SESSION['msg'];
+                unset($_SESSION['msg']);
+            }
+            ?>
+        </div>
 
     </main>
